split up schedulers into separate modules
